refactor(cymbal): move per-language raw frame structs into core/types
Relocate the langs module (all 13 raw frame structs, CommonFrameMetadata,
per-frame helpers, pure From<&RawX> conversions, native/utils) from
core/symbolication/langs to core/types/langs. The full network frame data
model — Frame, RawFrame, and every per-language raw frame — now lives in
core/types. crate::langs still resolves via the compat shim.

The catalog-backed resolve_frame bodies moved with their structs for now
and still reference the symbol-store Catalog (a residual core->symbolication
edge); the next step lifts that resolution behavior into symbolication so
core/types is pure data. cargo check + clippy green; langs tests pass.
