+++ rust/cymbal/src/core/symbolication/resolve.rs
@@ -3,12 +3,12 @@ use reqwest::Url;
 use crate::core::types::frames::{Frame, RawFrame};
 use crate::core::metric_consts::{FRAME_RESOLVED, LEGACY_JS_FRAME_RESOLVED, PER_FRAME_TIME};
 use crate::core::symbolication::symbol_store::Catalog;
-use crate::core::symbolication::langs::apple::RawAppleFrame;
-use crate::core::symbolication::langs::hermes::{HermesRef, RawHermesFrame};
-use crate::core::symbolication::langs::java::RawJavaFrame;
-use crate::core::symbolication::langs::js::RawJSFrame;
-use crate::core::symbolication::langs::native::DebugImage;
-use crate::core::symbolication::langs::node::RawNodeFrame;
+use crate::core::types::langs::apple::RawAppleFrame;
+use crate::core::types::langs::hermes::{HermesRef, RawHermesFrame};
+use crate::core::types::langs::java::RawJavaFrame;
+use crate::core::types::langs::js::RawJSFrame;
+use crate::core::types::langs::native::DebugImage;
+use crate::core::types::langs::node::RawNodeFrame;
 use crate::core::symbolication::symbol_store::apple::AppleRef;
 use crate::core::symbolication::symbol_store::chunk_id::OrChunkId;
 use crate::core::symbolication::symbol_store::hermesmap::ParsedHermesMap;
