+++ rust/cymbal/src/core/types/mod.rs
@@ -6,6 +6,7 @@ use serde::{Deserialize, Serialize};
 
 pub mod exception;
 pub mod frames;
+pub mod langs;
 pub mod stacktrace;
 
 pub use exception::*;
